Crate grafana_data_source

Grafana data source server

To display node statistics with Grafana, this module exposes a run_server function that starts up a HTTP server that conforms to the grafana-json-data-source API. The record_metrics macro can be used to pass metrics to this server.

Macros

record_metrics
Write metrics to METRICS.

Enums

Error
Error type that can be returned by either record_metrics or run_server.

Functions

record_metrics_slice
Write metrics to METRICS as a slice. Intended to be only used via record_metrics!.

run_server
Start the data source server.